Add a block quote
Broke up the first section of the first article with a block quote
pulled from a Washington Post article.  Used the cite attribute within
the block quote, and a separate cite tag to cover all the bases. The
Block quote includes a link to the original Washington Post article.
CSS background color and padding for the block quote.